// types/otherStyled.ts
export interface TimeFilterType {
  background?: string;
  color?: string;
  fontWeight?: string;
}

export interface OptionListType {
  left?: string;
}

export interface SelectType {
  background?: string;
  border?: string;
  paddingLeft?: string;
  color?: string;
  marginRight?: string;
  width?: string;
}

export interface TextType {
  size?: string;
  weight?: string | number;
  marginTop?: string;
  marginBottom?: string;
  marginLeft?: string;
}

export interface ButtonType {
  padding?: string;
  fontSize?: string;
  background?: string;
  margin?: string;
  hoverColor?: string;
}